// frontend-admin/src/api/generated/model/rksvReminderStatusDto.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Kasse API
 * Registrierkasse API - RKSV uyumlu kasa sistemi
 * OpenAPI spec version: v1
 */
import type { RksvReminderJahresbelegDto } from './rksvReminderJahresbelegDto';
import type { RksvReminderMonatsbelegDto } from './rksvReminderMonatsbelegDto';
import type { RksvReminderStartbelegDto } from './rksvReminderStartbelegDto';

export interface RksvReminderStatusDto {
  jahresbeleg: RksvReminderJahresbelegDto;
  monatsbeleg: RksvReminderMonatsbelegDto;
  startbeleg: RksvReminderStartbelegDto;
}